pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-ai-review"
version = "0.1.0"
description = "AI-powered code review tool using git diff and Gemini API"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
]

[project.scripts]
gar = "gar.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-mock>=3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gar*"]
